Worlds stored in unique files in worlds/ to begin supporting generation @ scale.
Renderer will also write images to `worlds/` directory using the same filename as the world, but add a
suffix describing the type of rendering (currently _terrain.png or _elevation.png). This should make it
fairly trivial to visualize a sample of maps by viewing the `worlds/` directory in any folder viewer
program.