Resolves #11, introduces world configuration file.
This commit extracts some key world generation variables (see #11 for list) that can now
be changes without digging through code. Three variables -- with others to be added as they
emerge -- can now be configured in world.ini, which will automatically be read during
generation.

Sane defaults are in place for all in case they're not specified.