Several significant performance and elevation quality improvements.
On the performance front, maps are very slow to generate when they are more than 1k cells.
For benchmarks where num_cells = 1024, the top two performance bottlenecks were
WorldRegion._build_graph() and WorldRegion.neighbors().

_build_graph() went from a .5755s runtime to 0.0089s per call (with one call per region), which
brought map generation times down from 8.86s to 5.61s. Importantly the new approach should also
scale better in larger maps.

neighbors() actually runs very quickly (0.00047s per call) but is called tens of millions of times,
so the challenge here is to reduce the number of times that the function needs to be called. I
accomplished this by caching some values that _form_plates() was re-calculating regularly that led
to a significant reduction in runtime. Map generation times were decreased from 5.61s to 3.01s,
including a reduction in _form_plates() from 2.633s to .987s.

Elevation quality improvements are based on adjustments to the noise generator, which is now easier
to adjust due to visual rendering of the noise field that's generated by a particular configuration.