Rudimentary city placement and brand new renderer!
New renderer is based on pycairo and is significantly faster and more flexible
than the pyplot-based renderer. Outputs are currently very similar to the legacy
renderer but the framework is in place make incremental improvements.

City placement is still basic but generates a score for each cell based on basic
definitions of survivability (can people survive there), economy (can people thrive
there), and threat (how risky is it to live there). Cities tend to occur near water
and far from each other.