Fixed a bug with voronoi region generation that sometimes causes a hot mess map to be generated.
Voronoi diagrams are generated iteratively using the centroid of each voronoi region as the point cloud
for the next iteration. In some cases, these centroids could be well outside of the [0, 1) region we
use for map generation; in rare cases many centroids would be out of this region and lead to a very sparse
voronoi diagram in the map region, which created the failure case leading to 'hot mess' maps.

This fixes the problem by checking that centroids are within [0, 1) before using them for successive
iterations. This does not guarantee that all centroids will be within this region, but it does limit
the percentage that can fall outside of the region and effectively eliminates the hot mess issue.