Fixed issue with temperature gen that was breaking biome generation.
Quality is still pretty mediocre, but at least it works as intended now! Temperature
is based on latitude and elevation (mostly the former). Confirmed that maps are
now being rendered more consistently now.

Additionally, glimpse now renders a png displaying the temperature and moisture
maps during world generation; these are the two conditions used for biome assignment,
so it should be easier to detect if things start going wrong with either in the
future.