Added bay detection.
This one took some work, but sets a foundation that is likely useful for other
POI detection. In short, glimpse can now detect bays (regions of water with
relatively high sinuosity) and define them as a named POI.

Quality isn't perfect but it looks pretty good, and a lot of the imperfections
aren't practically noticeable since the only rendering that's happening is the
name on top of an invisible set of cells.